Revert "Correciones iniciales."
This reverts commit 94d60445dde988ac0fa6865b9190833f186a55ce.
